example/backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
dynamic-rest-bse